requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.8
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
ruamel.yaml==0.17.22
simplejson==3.19.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15